Add support for Roblox App Beta for Windows
Roblox App Beta for Windows has two Lua States which share the same task scheduler. The beta menu state loads before the game does and creates a WaitingScriptJob. The game state will then load and create another WaitingScriptJob. Since the task scheduler is shared, the iterator will find the beta menu's WaitingScriptJob before the game's. The fix is to reverse the order of iteration so that the game's job is found first. This fix will work on both the default client and beta client.

To enable the beta for testing simply add a string value in Computer\HKEY_CURRENT_USER\SOFTWARE\ROBLOX Corporation\Environments\roblox-player
Create a string value named "LaunchExp" with the value "InApp" then start the game

Info about the beta:
https://devforum.roblox.com/t/psa-the-roblox-app-beta-for-windows-is-now-underway/925069

It's very possible that this code doesn't work because I couldn't get the project to compile with vs 2019 and couldn't test it.